feat: Add LimitWarning, DurationField, QuickSelectButton components and enhance FormField with dynamic ID and iOS support
